png-renderer/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.25.0
playwright==1.41.0
python-multipart==0.0.7
pydantic==2.5.2
pydantic-settings==2.1.0
pillow==10.2.0
python-dotenv==1.0.0
httpx==0.25.2
loguru==0.7.2